fix public catalog detail page (auth gate + HTTPS self-fetch 404)
Two bugs on /catalog/[id]:

a) It required authentication. proxy.ts only allowlisted the exact path
   /catalog, so /catalog/<id> was gated and redirected to sign-in. Allow
   /catalog/ prefixed detail pages too.

b) When authenticated it 404'd. getServerSideProps fetched its own API over
   http://<host>, which fails on an HTTPS deployment (http request to an
   https port), and the page returns notFound on any failure. Read the DB
   directly instead via a shared getPublicBookDetail(), which also drops a
   needless self-HTTP round-trip. The API route uses the same shared query.

The public book shape lives in one place (PUBLIC_BOOK_SELECT and
toPublicBook, reused by the list endpoint, the related-books block and the
detail mapping), the related-books candidate query is bounded with take: 200,
parseTopics is a shared export instead of a per-file copy, and the page runs
the same dev-only reconnectPrisma() as the API route.

Verified: /catalog/<id> loads unauthenticated (200), a missing id 404s.

Fixes #447
